feat: implementação inicial do fluentvalidation e UnitOfWork

+++ ToDo.Domain/Commands/UsuarioCommands/UsuarioCommandHandler.cs
@@ -1,65 +1,76 @@
-﻿using MediatR;
+﻿using FluentValidation.Results;
+using MediatR;
+using ToDo.Domain.Core.Interfaces;
 using ToDo.Domain.Interfaces;
 using ToDo.Domain.Models;
 
 namespace ToDo.Domain.Commands.UsuarioCommands
 {
-    public class UsuarioCommandHandler : 
-        IRequestHandler<AddUserCommand, bool>,
-        IRequestHandler<EditUserCommand, bool>,
-        IRequestHandler<DeleteUserCommand, bool>
+    public class UsuarioCommandHandler : BaseCommandHandler,
+        IRequestHandler<AddUserCommand, ValidationResult>,
+        IRequestHandler<EditUserCommand, ValidationResult>,
+        IRequestHandler<DeleteUserCommand, ValidationResult>
     {
 
         private readonly IUsuarioRepository _usuarioRepository;
 
-        public UsuarioCommandHandler(IUsuarioRepository usuarioRepository)
+        public UsuarioCommandHandler(IUsuarioRepository usuarioRepository, IUnitOfWork uow) : base(uow)
         {
             _usuarioRepository = usuarioRepository;
         }
 
-        public async Task<bool> Handle(AddUserCommand request, CancellationToken cancellationToken)
+        public async Task<ValidationResult> Handle(AddUserCommand request, CancellationToken cancellationToken)
         {
-            //valida se dados usuário válido
+            if (!request.IsValid()) return request.ValidationResult;
 
             var existingUser = await _usuarioRepository.GetByNameAsync(request.Name);
 
             if (existingUser != null) {
-                return false;
+                AddError("Já existe um usuário com o nome informado.");
+                return _validationResult;
             }
 
             var newUser = new Usuario(request.Id, request.Name);
             _usuarioRepository.Add(newUser);
-            return await _usuarioRepository.SaveChangesAsync();
+            await CommitAsync();
+
+            return _validationResult;
         }
 
-        public async Task<bool> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+        public async Task<ValidationResult> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
-            //valida se dados usuário válido
+            if (!request.IsValid()) return request.ValidationResult;
 
             var existingUser = await _usuarioRepository.GetByIdAsync(request.Id);
 
             if (existingUser is null) {
-                return false;
+                AddError("Usuário não encontrado.");
+                return _validationResult;
             }
 
             _usuarioRepository.Delete(existingUser);
-            return await _usuarioRepository.SaveChangesAsync();
+            await CommitAsync();
+
+            return _validationResult;
         }
 
-        public async Task<bool> Handle(EditUserCommand request, CancellationToken cancellationToken)
+        public async Task<ValidationResult> Handle(EditUserCommand request, CancellationToken cancellationToken)
         {
-            //valida se dados usuário válido
+            if (!request.IsValid()) return request.ValidationResult;
 
             var existingUser = await _usuarioRepository.GetByNameAsync(request.Name);
 
             if (existingUser != null && existingUser.Id != request.Id)
             {
-                return false;
+                AddError("Foram encontradas inconsistências nas informações.");
+                return _validationResult;
             }
 
             var newUser = new Usuario(request.Id, request.Name);
             _usuarioRepository.Update(newUser);
-            return await _usuarioRepository.SaveChangesAsync();
+            await CommitAsync();
+
+            return _validationResult;
         }
     }
 }
